Refactor the routes for better DX. Also rename manage and data view files for easier mental model.
