+++ src/app/(private)/dashboard/(staff)/(data-view)/inventory/_client.tsx
