+++ src/app/(private)/dashboard/(staff)/(data-view)/inventory/page.tsx
