+++ src/app/(private)/dashboard/(staff)/(data-view)/loading.tsx
@@ -1,0 +1,5 @@
+import { DataViewLoadingClient } from "@/components/manage/data-view-loading";
+
+export default function StaffDataViewLoading() {
+  return <DataViewLoadingClient />;
+}
