+++ src/app/(private)/dashboard/(staff)/layout.tsx
@@ -2,7 +2,7 @@ import { getUserProfile } from "@/lib/auth-queries";
 import { redirect } from "next/navigation";
 import { ReactNode } from "react";
 
-export default async function ManageLayout({
+export default async function StaffLayout({
   children,
 }: {
   children: ReactNode;
